+++ src/Requests/Assistants/File/ListAssistantFiles.php
@@ -1,30 +1,30 @@
 <?php
 
-namespace ChrisReedIO\OpenAI\SDK\Requests\Assistants;
+namespace ChrisReedIO\OpenAI\SDK\Requests\Assistants\File;
 
 use Saloon\Enums\Method;
 use Saloon\Http\Request;
 
 /**
- * listRuns
+ * listAssistantFiles
  */
-class ListRuns extends Request
+class ListAssistantFiles extends Request
 {
     protected Method $method = Method::GET;
 
     public function resolveEndpoint(): string
     {
-        return "/threads/{$this->threadId}/runs";
+        return "/assistants/{$this->assistantId}/files";
     }
 
     /**
-     * @param  string  $threadId The ID of the thread the run belongs to.
+     * @param  string  $assistantId The ID of the assistant the file belongs to.
      * @param  null|int  $limit A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 20.
      * @param  null|string  $order Sort order by the `created_at` timestamp of the objects. `asc` for ascending order and `desc` for descending order.
      * @param  null|string  $before A cursor for use in pagination. `before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can include before=obj_foo in order to fetch the previous page of the list.
      */
     public function __construct(
-        protected string $threadId,
+        protected string $assistantId,
         protected ?int $limit = null,
         protected ?string $order = null,
         protected ?string $before = null,
