+++ src/Requests/Messages/File/GetMessageFile.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChrisReedIO\OpenAI\SDK\Requests\Assistants;
+namespace ChrisReedIO\OpenAI\SDK\Requests\Messages\File;
 
 use Saloon\Enums\Method;
 use Saloon\Http\Request;
