+++ src/Requests/Runs/CreateRun.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChrisReedIO\OpenAI\SDK\Requests\Assistants;
+namespace ChrisReedIO\OpenAI\SDK\Requests\Runs;
 
 use Saloon\Contracts\Body\HasBody;
 use Saloon\Enums\Method;
@@ -8,9 +8,9 @@ use Saloon\Http\Request;
 use Saloon\Traits\Body\HasJsonBody;
 
 /**
- * modifyThread
+ * createRun
  */
-class ModifyThread extends Request implements HasBody
+class CreateRun extends Request implements HasBody
 {
     use HasJsonBody;
 
@@ -18,11 +18,11 @@ class ModifyThread extends Request implements HasBody
 
     public function resolveEndpoint(): string
     {
-        return "/threads/{$this->threadId}";
+        return "/threads/{$this->threadId}/runs";
     }
 
     /**
-     * @param  string  $threadId The ID of the thread to modify. Only the `metadata` can be modified.
+     * @param  string  $threadId The ID of the thread to run.
      */
     public function __construct(
         protected string $threadId,
