+++ src/Requests/Threads/ModifyThread.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChrisReedIO\OpenAI\SDK\Requests\Assistants;
+namespace ChrisReedIO\OpenAI\SDK\Requests\Threads;
 
 use Saloon\Contracts\Body\HasBody;
 use Saloon\Enums\Method;
@@ -8,9 +8,9 @@ use Saloon\Http\Request;
 use Saloon\Traits\Body\HasJsonBody;
 
 /**
- * createRun
+ * modifyThread
  */
-class CreateRun extends Request implements HasBody
+class ModifyThread extends Request implements HasBody
 {
     use HasJsonBody;
 
@@ -18,11 +18,11 @@ class CreateRun extends Request implements HasBody
 
     public function resolveEndpoint(): string
     {
-        return "/threads/{$this->threadId}/runs";
+        return "/threads/{$this->threadId}";
     }
 
     /**
-     * @param  string  $threadId The ID of the thread to run.
+     * @param  string  $threadId The ID of the thread to modify. Only the `metadata` can be modified.
      */
     public function __construct(
         protected string $threadId,
