+++ src/Resources/Assistants.php
@@ -2,34 +2,34 @@
 
 namespace ChrisReedIO\OpenAI\SDK\Resources;
 
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\CancelRun;
 use ChrisReedIO\OpenAI\SDK\Requests\Assistants\CreateAssistant;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\CreateAssistantFile;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\CreateMessage;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\CreateRun;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\CreateThread;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\CreateThreadAndRun;
 use ChrisReedIO\OpenAI\SDK\Requests\Assistants\DeleteAssistant;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\DeleteAssistantFile;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\DeleteThread;
+use ChrisReedIO\OpenAI\SDK\Requests\Assistants\File\CreateAssistantFile;
+use ChrisReedIO\OpenAI\SDK\Requests\Assistants\File\DeleteAssistantFile;
+use ChrisReedIO\OpenAI\SDK\Requests\Assistants\File\GetAssistantFile;
+use ChrisReedIO\OpenAI\SDK\Requests\Assistants\File\ListAssistantFiles;
 use ChrisReedIO\OpenAI\SDK\Requests\Assistants\GetAssistant;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\GetAssistantFile;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\GetMessage;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\GetMessageFile;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\GetRun;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\GetRunStep;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\GetThread;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ListAssistantFiles;
 use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ListAssistants;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ListMessageFiles;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ListMessages;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ListRuns;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ListRunSteps;
 use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ModifyAssistant;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ModifyMessage;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ModifyRun;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\ModifyThread;
-use ChrisReedIO\OpenAI\SDK\Requests\Assistants\SubmitToolOuputsToRun;
+use ChrisReedIO\OpenAI\SDK\Requests\Messages\CreateMessage;
+use ChrisReedIO\OpenAI\SDK\Requests\Messages\File\GetMessageFile;
+use ChrisReedIO\OpenAI\SDK\Requests\Messages\File\ListMessageFiles;
+use ChrisReedIO\OpenAI\SDK\Requests\Messages\GetMessage;
+use ChrisReedIO\OpenAI\SDK\Requests\Messages\ListMessages;
+use ChrisReedIO\OpenAI\SDK\Requests\Messages\ModifyMessage;
+use ChrisReedIO\OpenAI\SDK\Requests\Runs\CancelRun;
+use ChrisReedIO\OpenAI\SDK\Requests\Runs\CreateRun;
+use ChrisReedIO\OpenAI\SDK\Requests\Runs\GetRun;
+use ChrisReedIO\OpenAI\SDK\Requests\Runs\ListRuns;
+use ChrisReedIO\OpenAI\SDK\Requests\Runs\ModifyRun;
+use ChrisReedIO\OpenAI\SDK\Requests\Runs\SubmitToolOuputsToRun;
+use ChrisReedIO\OpenAI\SDK\Requests\RunSteps\GetRunStep;
+use ChrisReedIO\OpenAI\SDK\Requests\RunSteps\ListRunSteps;
+use ChrisReedIO\OpenAI\SDK\Requests\Threads\CreateThread;
+use ChrisReedIO\OpenAI\SDK\Requests\Threads\CreateThreadAndRun;
+use ChrisReedIO\OpenAI\SDK\Requests\Threads\DeleteThread;
+use ChrisReedIO\OpenAI\SDK\Requests\Threads\GetThread;
+use ChrisReedIO\OpenAI\SDK\Requests\Threads\ModifyThread;
 use ReflectionException;
 use Saloon\Http\Response;
 use Throwable;
