+++ src/Resources/RunSteps.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ChrisReedIO\OpenAI\SDK\Resources;
+
+class RunSteps extends Resource
+{
+
+}
