+++ src/Resources/Threads.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ChrisReedIO\OpenAI\SDK\Resources;
+
+class Threads extends Resource
+{
+
+}
